# Optimise ShardAwareDeduplicateFilter