--- a/pkg/compactor/shard_aware_deduplicate_filter_test.go
+++ b/pkg/compactor/shard_aware_deduplicate_filter_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	promtest "github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/prometheus/prometheus/tsdb"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	mimir_tsdb "github.com/grafana/mimir/pkg/storage/tsdb"
@@ -460,28 +459,6 @@ func BenchmarkDeduplicateFilter_Filter(b *testing.B) {
 			}
 		})
 	}
-}
-
-func TestNewBlockWithSuccessors(t *testing.T) {
-	t.Run("source block IDs should be sorted by timestamp", func(t *testing.T) {
-		b := newBlockWithSuccessors(&block.Meta{
-			BlockMeta: tsdb.BlockMeta{
-				Compaction: tsdb.BlockMetaCompaction{
-					Sources: []ulid.ULID{
-						ulid.MustNew(10, nil),
-						ulid.MustNew(30, nil),
-						ulid.MustNew(20, nil),
-					},
-				},
-			},
-		})
-
-		assert.Equal(t, []ulid.ULID{
-			ulid.MustNew(10, nil),
-			ulid.MustNew(20, nil),
-			ulid.MustNew(30, nil),
-		}, b.sourcesSortedList)
-	})
 }
 
 func TestBloomFilter(t *testing.T) {