#### What this PR does

This PR has been be authored by @colega and myself.

We had an issue where a tenant compaction was lagging behind, and their number blocks reached the monster number of 250K (from a baseline of 10K). When that happened, we found out that the compactor planning was extremely slow (3 hours), and that the slow down came from `ShardAwareDeduplicateFilter`.

In this PR we introduce some small optimizations to `ShardAwareDeduplicateFilter` for tenants with a very large number of blocks. The impact on performance is not impressive, but we got a -33% reduction (from 3h30m to 2h15m) for the real world scenario.

Existing benchmark result:

```

goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/compactor
cpu: Apple M3 Pro
                                            │  before.txt   │ after-with-bloom-filter-and-without-list.txt │
                                            │    sec/op     │        sec/op          vs base               │
DeduplicateFilter_Filter/Block-10/#00-11       17.64µ ± 14%             25.23µ ± 3%  +43.04% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10/#01-11       53.33µ ±  1%             76.36µ ± 1%  +43.18% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-100/#00-11      383.6µ ±  4%             294.2µ ± 1%  -23.30% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-100/#01-11     1153.2µ ±  1%             882.1µ ± 1%  -23.51% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-1000/#00-11    28.346m ±  1%             7.578m ± 6%  -73.26% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-1000/#01-11     90.15m ±  1%             23.02m ± 1%  -74.46% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10000/#00-11     6.820 ±  5%              1.237 ± 1%  -81.86% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10000/#01-11    20.710 ±  2%              3.723 ± 1%  -82.02% (p=0.002 n=6)
geomean                                        10.52m                   5.018m       -52.31%

                                            │  before.txt  │ after-with-bloom-filter-and-without-list.txt │
                                            │     B/op     │         B/op           vs base               │
DeduplicateFilter_Filter/Block-10/#00-11      24.87Ki ± 0%            30.59Ki ± 0%  +23.00% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10/#01-11      74.61Ki ± 0%            91.76Ki ± 0%  +23.00% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-100/#00-11     247.4Ki ± 0%            299.5Ki ± 0%  +21.07% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-100/#01-11     742.6Ki ± 0%            898.9Ki ± 0%  +21.05% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-1000/#00-11    2.509Mi ± 0%            2.971Mi ± 0%  +18.40% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-1000/#01-11    7.955Mi ± 0%            9.033Mi ± 0%  +13.54% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10000/#00-11   48.80Mi ± 0%            58.87Mi ± 0%  +20.64% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10000/#01-11   146.4Mi ± 0%            176.6Mi ± 0%  +20.64% (p=0.002 n=6)
geomean                                       1.606Mi                 1.929Mi       +20.13%

                                            │ before.txt  │ after-with-bloom-filter-and-without-list.txt │
                                            │  allocs/op  │       allocs/op         vs base              │
DeduplicateFilter_Filter/Block-10/#00-11       63.00 ± 0%               64.00 ± 0%  +1.59% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10/#01-11       189.0 ± 0%               192.0 ± 0%  +1.59% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-100/#00-11      522.0 ± 0%               523.0 ± 0%  +0.19% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-100/#01-11     1.567k ± 0%              1.570k ± 0%  +0.19% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-1000/#00-11    5.162k ± 0%              5.069k ± 0%  -1.80% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-1000/#01-11    16.37k ± 0%              15.41k ± 0%  -5.83% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10000/#00-11   100.2k ± 0%              100.2k ± 0%  +0.00% (p=0.002 n=6)
DeduplicateFilter_Filter/Block-10000/#01-11   300.6k ± 0%              300.6k ± 0%  +0.00% (p=0.002 n=6)
geomean                                       3.542k                   3.523k       -0.53%
```

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry is not needed, please add the `changelog-not-needed` label to the PR.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.